pyproject.toml
[project]
name = "migration-engine"
version = "0.1.0"
description = "Multi-agent migration feasibility and risk engine for US citizens considering relocation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "google-generativeai>=0.8.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dateutil>=2.9.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["migration_engine"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]
